[VM] Remove compiler warinings. [General] Update SOVERSION. [COMMON] Fix buffer overflow around strncpy().

+++ source/src/vm/ym2413.cpp
@@ -693,7 +693,7 @@ MAME_INLINE void advance(YM2413C *chip)
         {
           op->volume += eg_inc[op->eg_sel_dr + ((chip->eg_cnt>>op->eg_sh_dr)&7)];
 
-          if ( op->volume >= op->sl )
+          if ( op->volume >= (INT32)(op->sl) )
             op->state = EG_SUS;
         }
       break;
@@ -1749,7 +1749,7 @@ static void OPLLWriteReg(YM2413C *chip, int r, int v)
       CH->sus = v & 0x20;
     }
     /* update */
-    if(CH->block_fnum != block_fnum)
+    if(CH->block_fnum != (UINT32)block_fnum)
     {
       UINT8 block;
